requirements.txt
absl-py==1.2.0
aicrowd-gym==0.0.3
antlr4-python3-runtime==4.8
asttokens==2.0.5
atari-py==0.2.6
attrs==21.4.0
backcall==0.2.0
box2d-py==2.3.8
cachetools==4.2.4
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==1.3.0
coloredlogs==15.0.1
cycler==0.11.0
Cython==0.29.30
daemoniker==0.2.3
decorator==5.1.1
dill==0.3.5.1
docker-pycreds==0.4.0
executing==0.9.1
fasteners==0.17.3
flaky==3.7.0
fonttools==4.34.4
future==0.18.2
getch==1.0
gitdb==4.0.9
GitPython==3.1.27
glcontext==2.3.6
glfw==1.12.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
gym==0.19.0
gym-minigrid==1.0.2
gym3==0.3.3
humanfriendly==10.0
hydra-core==1.0.6
idna==3.3
ImageHash==4.2.1
imageio==2.19.5
imageio-ffmpeg==0.3.0
importlib-metadata==4.12.0
importlib-resources==5.8.0
inflection==0.5.1
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
kiwisolver==1.4.4
lockfile==0.12.2
loguru==0.6.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
minerl==1.0.0
minio==7.1.11
moderngl==5.6.4
msgpack==1.0.4
msgpack-numpy==0.4.8
mujoco-py==2.1.2.14
numpy==1.23.1
oauthlib==3.2.0
omegaconf==2.0.6
opencv-python==4.5.1.48
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyglet==1.5.0
Pygments==2.12.0
pyparsing==3.0.9
Pyro4==4.82
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.8.1
sentry-sdk==1.8.0
serpent==1.41
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
stable-baselines3==1.0
stack-data==0.3.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.1
termcolor==1.1.0
timeout-decorator==0.5.0
torch==1.9.0
torchvision==0.8.2
tornado==5.1.1
tqdm==4.42.1
traitlets==5.3.0
typing==3.7.4.3
typing-extensions==4.3.0
urllib3==1.26.10
wandb==0.12.21
wcwidth==0.2.5
Werkzeug==2.1.2
xmltodict==0.12.0
zipp==3.8.1
zmq==0.0.0
